気象情報関連のクラスを別ファイルに切り出し
